# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
      "ApiErrorResponseDto": {
        "properties": {
          "code": {
            "enum": [
              "ACCESS_DENIED",
              "AUTHENTICATION_REQUIRED",
              "INTERNAL_ERROR",
              "REQUEST_INVALID",
              "RESOURCE_NOT_FOUND",
              "SERVICE_UNAVAILABLE",
              "FORBIDDEN_CAPABILITY",
              "PAYEE_FORBIDDEN",
              "SELLER_IDENTITY_FORBIDDEN",
              "SINGLE_MERCHANT_VIOLATION",
              "ACTOR_SPOOFED",
              "ACCOUNT_TYPE_INVALID",
              "APPROVAL_VERSION_CONFLICT",
              "DATA_SCOPE_FORBIDDEN",
              "FIELD_FORBIDDEN",
              "IDEMPOTENCY_CONFLICT",
              "SECOND_VERIFICATION_REQUIRED",
              "STATE_TRANSITION_INVALID",
              "SUPPLIER_DUPLICATE",
              "SUPPLIER_SCOPE_FORBIDDEN",
              "VALIDATION_FAILED",
              "VERSION_CONFLICT",
              "WORKSPACE_FORBIDDEN",
              "ACCOUNT_SUSPENDED",
              "AUTH_INVALID",
              "AUTH_SESSION_REVOKED",
              "RATE_LIMITED",
              "SUPPLIER_NOT_ACTIVE",
              "WORKSPACE_MENU_VIOLATION",
              "WORKSPACE_MODULE_NOT_FOUND",
              "WORKSPACE_SELECTION_REQUIRED",
              "WORKSPACE_SESSION_CONFLICT",
              "AUDIT_IMMUTABLE",
              "AUDIT_REQUIRED",
              "EXPORT_APPROVAL_REQUIRED",
              "REQUEST_ID_REQUIRED",
              "SAME_NATURAL_PERSON_REVIEW",
              "SECOND_REVIEW_REQUIRED",
              "APPROVAL_NOT_FOUND",
              "APPROVAL_STATE_INVALID",
              "IDEMPOTENCY_KEY_CONFLICT",
              "IDEMPOTENCY_KEY_REQUIRED",
              "APPAREL_HISTORY_REWRITE",
              "APPAREL_REQUIRED_FIELD_MISSING",
              "BUNDLE_SCHEMA_INVALID",
              "DIGITAL_HISTORY_REWRITE",
              "DIGITAL_MODEL_DUPLICATE",
              "DIGITAL_REQUIRED_FIELD_MISSING",
              "CATEGORY_DISABLED",
              "CATEGORY_DUPLICATE",
              "CATEGORY_LEVEL_INVALID",
              "CATEGORY_NOT_FOUND",
              "CATEGORY_NOT_LEAF",
              "CATEGORY_PARENT_INVALID",
              "CATEGORY_REFERENCED",
              "CATEGORY_TEMPLATE_INVALID",
              "PRICE_FIELD_FORBIDDEN",
              "PRICE_INVALID",
              "INITIAL_PRICE_REVIEW_PENDING",
              "INITIAL_PRICE_STATE_INVALID",
              "PRODUCT_APPROVAL_INCOMPLETE",
              "PRODUCT_NOT_FOUND",
              "PRODUCT_NOT_SALEABLE",
              "SUPPLIER_INACTIVE",
              "SUPPLIER_PRODUCT_DUPLICATE",
              "SUPPLIER_PRODUCT_NOT_FOUND",
              "SUPPLIER_SKU_DUPLICATE",
              "SELF_APPROVAL_FORBIDDEN",
              "SKU_DIMENSION_DUPLICATE",
              "FRESH_HISTORY_REWRITE",
              "FRESH_REQUIRED_FIELD_MISSING",
              "FRESH_WEIGHT_RULE_INVALID",
              "REGULATORY_WARNING_REQUIRED",
              "REGULATED_CATEGORY_DISABLED",
              "QUALIFICATION_REQUIRED",
              "TEMPLATE_DATA_INVALID",
              "TEMPLATE_DRAFT_EXISTS",
              "TEMPLATE_IMMUTABLE",
              "TEMPLATE_NOT_FOUND",
              "TEMPLATE_SCHEMA_INVALID",
              "TEMPLATE_VERSION_INACTIVE",
              "TEMPLATE_VERSION_IMMUTABLE"
            ],
            "example": "RESOURCE_NOT_FOUND",
            "type": "string"
          },
          "message": {
            "example": "Resource was not found",
            "type": "string"
          },
          "path": {
            "example": "/missing",
            "type": "string"
          },
          "requestId": {
            "example": "contract-request-0001",
            "type": "string"
          },
          "statusCode": {
            "example": 404,
            "type": "number"
          },
          "timestamp": {
            "example": "2026-08-02T00:00:00.000Z",
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "statusCode",
          "code",
          "message",
          "requestId",
          "path",
          "timestamp"
        ],
        "type": "object"
      },
...
      "RegulatedCategoryControlPageDto": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RegulatedCategoryControlResponseDto"
            },
            "type": "array"
          },
          "total": {
            "minimum": 0,
            "type": "number"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "type": "object"
      },
      "RegulatedCategoryControlResponseDto": {
        "properties": {
          "categoryId": {
            "format": "uuid",
            "type": "string"
          },
          "companyQualificationReferenceCount": {
            "minimum": 0,
            "type": "number"
          },
          "disabledAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "enabledAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "qualificationValidUntil": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "status": {
            "enum": [
              "DISABLED",
              "ENABLED"
            ],
            "type": "string"
          },
          "version": {
            "minimum": 0,
            "type": "number"
          }
        },
        "required": [
          "id",
          "categoryId",
          "status",
          "companyQualificationReferenceCount",
          "qualificationValidUntil",
          "version"
        ],
        "type": "object"
      },
      "RegulatedCategoryDisableRequestDto": {
        "properties": {
          "reason": {
            "maxLength": 500,
            "minLength": 2,
            "type": "string"
          },
          "secondVerificationCode": {
            "maxLength": 64,
            "minLength": 4,
            "type": "string"
          },
          "version": {
            "minimum": 0,
            "type": "number"
          }
        },
        "required": [
          "version",
          "reason",
          "secondVerificationCode"
        ],
        "type": "object"
      },
      "RegulatedCategoryEnableRequestDto": {
        "properties": {
          "companyQualificationReferences": {
            "items": {
              "type": "string"
            },
            "maxItems": 50,
            "minItems": 1,
            "type": "array"
          },
          "qualificationValidUntil": {
            "format": "date-time",
            "type": "string"
          },
          "secondVerificationCode": {
            "maxLength": 64,
            "minLength": 4,
            "type": "string"
          },
          "version": {
            "minimum": 0,
            "type": "number"
          }
        },
        "required": [
          "version",
          "companyQualificationReferences",
          "qualificationValidUntil",
          "secondVerificationCode"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "companyFunctionalSession": {
        "description": "Secure HttpOnly company functional account session",
        "in": "cookie",
        "name": "__Host-fulishe-company-admin",
        "type": "apiKey"
      },
      "functionalSession": {
        "description": "Server-bound functional account session; clients cannot select owner scope",
        "in": "cookie",
        "name": "fulishe_session",
        "type": "apiKey"
      },
      "supplierFunctionalSession": {
        "description": "Secure HttpOnly supplier functional account session",
        "in": "cookie",
        "name": "__Host-fulishe-supplier-portal",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "江苏福礼团供应链科技有限公司单商户平台 API 契约",
    "title": "福礼社统一 API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/company/regulated-category-controls": {
      "get": {
        "operationId": "regulatedCategoryControls.list",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegulatedCategoryControlPageDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "List company-scoped high-risk controls",
        "tags": [
          "company-regulated-category-controls"
        ]
      }
    },
    "/v1/company/regulated-category-controls/{categoryId}/disable": {
      "post": {
        "operationId": "regulatedCategoryControls.disable",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "categoryId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegulatedCategoryDisableRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegulatedCategoryControlResponseDto"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "Disable a high-risk category after second verification",
        "tags": [
          "company-regulated-category-controls"
        ]
      }
    },
    "/v1/company/regulated-category-controls/{categoryId}/enable": {
      "post": {
        "operationId": "regulatedCategoryControls.enable",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "categoryId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegulatedCategoryEnableRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegulatedCategoryControlResponseDto"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "Enable a high-risk category after second verification",
        "tags": [
          "company-regulated-category-controls"
        ]
      }
    },
...
  },
  "servers": [],
...
}